+++ changelog.md
@@ -3,6 +3,14 @@ All notable changes to this project will be documented in this file.
 
 The format is loosely based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.5.36] - 2020-03-20
+
+- Added new page, [Vietnam War Photographs from the Martin Cavazos Center](https://codyburleson.com/photos/cavazos-center)
+- Added new page,  [U.S. Marine Corps Platoon 156 - San Diego, 1966](https://codyburleson.com/marine-platoon-156-san-diego-1966)
+- Modified content of [/about](https://codyburleson.com/about) page.
+- Updated `@ionic/core` to  `^5.0.5`
+- Removed unecessary `ionicons` package dependency.
+
 ## [1.5.35] - 2020-03-03
 
 - Added new post: [http://localhost:3333/unreal-engine-doc-expanded](https://codyburleson.com/http://localhost:3333/unreal-engine-doc-expanded)